student components completed

+++ CurriculumManagementApp/src/app/components/headmaster/student/view-students/view-students.component.ts
@@ -17,8 +17,9 @@ export class ViewStudentsComponent implements OnInit {
   public student: Student = new Student();
   public standardList: string[] = ['I', 'II', 'III', 'IV', 'V', 'VI', 'VII', 'VIII', 'IX', 'X', 'XI', 'XII'];
   public classList: Class[] = [];
-  public roomNo: number =0;
-  public isShown: boolean = false;
+  public roomNo: number = 0;
+  public isHidden: boolean = false;
+  public errorMessage: string = "";
   constructor(private studentService: StudentService,
     private classService: ClassService,
     private router: Router) { }
@@ -32,36 +33,44 @@ export class ViewStudentsComponent implements OnInit {
   }
   getSections() {
     this.classService.getClassesByStandard(this.standard?.value).subscribe(response => {
-      let responseBody:Response=response;
+      let responseBody: Response = response;
       this.classList = responseBody.data;
       console.log(this.classList);
-    },error=>{
+    }, error => {
       window.alert(error.error.message);
-    })
+    });
   }
   getStudents() {
     this.classService.getClassRoomNo(this.standard?.value, this.section?.value).subscribe(response => {
-      let responseBody:Response = response;
+      let responseBody: Response = response;
       console.log(responseBody.data);
       this.roomNo = responseBody.data;
       this.studentService.getStudents(this.roomNo).subscribe(response => {
-        let responseBody:Response = response;
+        let responseBody: Response = response;
         console.log(responseBody.data);
         this.studentsList = responseBody.data;
-      },error=>{
+        this.isHidden = false;
+      }, error => {
+        this.errorMessage = error.error.message;
+        this.isHidden = true;
         window.alert(error.error.message);
       });
-    },error=>{
+    }, error => {
+      this.errorMessage = error.error.message;
       window.alert(error.error.message);
     });
-    this.isShown = true;
   }
   deleteStudent() {
-    this.studentService.deleteStudent(this.ViewStudentForm.get('option')?.value).subscribe(response => {
-      let responseBody:Response=response;
-      window.alert(responseBody.message);
-      this.getStudents;
-    });
+    let response: boolean = window.confirm("Are you sure want to continue?");
+    if (response) {
+      this.studentService.deleteStudent(this.ViewStudentForm.get('option')?.value).subscribe(response => {
+        let responseBody: Response = response;
+        window.alert(responseBody.message);
+        this.getStudents;
+      }, error => {
+        window.alert(error.error.message);
+      });
+    }
   }
   backToMain() {
     this.router.navigate(['admin'])
